fix transformTemplate; fix visualElement hydration
